+++ non_catalog_apps/prog_calculator/calculation_logic.h
@@ -8,7 +8,7 @@ bool decToHex(const char* decString, char* decToHexResult, size_t resultSize);
 bool decToChar(const char* decString, char* outChar);
 bool hexToBin(const char* hexString, char* binaryResult, size_t resultSize);
 bool hexToDec(const char* hexString, unsigned long long* outNum);
-bool binToDec(const char* binaryString, int* decResult);
+bool binToDec(const char* binaryString, unsigned long long* decResult);
 bool binToHex(const char* binaryString, char* hexResult, size_t resultSize);
 void calculate(Calculator* calculator_state);
